% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_get_file.R
\name{entsoe_get_file}
\alias{entsoe_get_file}
\title{Function to get the file for a basis_name, year and month.}
\usage{
entsoe_get_file(basis_name = NULL, year = lubridate::year(Sys.Date()),
  month = lubridate::month(Sys.Date()), file_name = NULL)
}
\arguments{
\item{basis_name}{High level folder name of the file.}

\item{year}{Year.}

\item{month}{Month (example: 1 or 11).}

\item{file_name}{File name of file to download. If this param is used it will overwrite basis_name, year and month.}
}
\description{
Function to get the file for a basis_name, year and month.
}
