% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_get_log.R
\name{entsoe_get_log}
\alias{entsoe_get_log}
\title{Function to get the file for a basis_name, year and month.}
\usage{
entsoe_get_log()
}
\arguments{
\item{basis_name}{High level folder name of the file.}

\item{year}{Year.}

\item{month}{Month (example: 1 or 11).}
}
\description{
Function to get the file for a basis_name, year and month.
}
