% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_list_files.R
\name{entsoe_list_files}
\alias{entsoe_list_files}
\title{Function to list files in a directory.}
\usage{
entsoe_list_files(basis_name)
}
\arguments{
\item{basis_name}{High level folder name of the file.}
}
\description{
Function to list files in a directory.
}
