% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_list_folders.R
\name{entsoe_list_folders}
\alias{entsoe_list_folders}
\title{List top level names of links to the low level folders.}
\usage{
entsoe_list_folders()
}
\description{
List top level names of links to the low level folders.
}
