% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_get_file.R
\name{entsoe_split_file_name}
\alias{entsoe_split_file_name}
\title{Split a filename into it's year, month and basis_name components.}
\usage{
entsoe_split_file_name(file_name)
}
\arguments{
\item{file_name}{File name.}
}
\description{
Split a filename into it's year, month and basis_name components.
}
