% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_list_files.R
\name{parse_files}
\alias{parse_files}
\title{Thanks to Scott Chamberlain (from the ropensci/ftp package).}
\usage{
parse_files(x)
}
\description{
Thanks to Scott Chamberlain (from the ropensci/ftp package).
}
