% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_list_folders.R
\name{parse_links}
\alias{parse_links}
\title{Thanks to Scott Chamberlain (from the ropensci/ftp package).}
\usage{
parse_links(x)
}
\description{
Thanks to Scott Chamberlain (from the ropensci/ftp package).
}
